API DTO Updates

+++ transaction-api/DTOs/ClientTransactionDTO.cs
@@ -2,7 +2,7 @@
 {
     public class ClientTransactionDTO
     {
-        public int TransactionID { get; set; }
+        public long TransactionID { get; set; }
         public string TransactionTypeName { get; set; }
         public string Comment { get; set; }
         public int TransactionTypeID { get; set; }
